Updated thomas to handle general matrices and removed old cudss binding + added some GPU syncronization
