+++ src/qttools/wave_function_solver/__init__.py
@@ -2,7 +2,6 @@
 
 from qttools import sparse
 from qttools.wave_function_solver.cudss import cuDSS
-from qttools.wave_function_solver.cudss_v2 import cudss_solver_2
 from qttools.wave_function_solver.mumps import MUMPS
 from qttools.wave_function_solver.pardiso import PARDISO
 from qttools.wave_function_solver.solver import WFSolver
@@ -15,7 +14,6 @@ preferred_matrix_type = {
     "cudss": sparse.csr_matrix,
     "pardiso": sparse.csr_matrix,
     "thomas": sparse.csr_matrix,
-    "cudss_v2": sparse.csr_matrix,
 }
 
 __all__ = [
@@ -26,5 +24,4 @@ __all__ = [
     "PARDISO",
     "Thomas",
     "preferred_matrix_type",
-    "cudss_solver_2",
 ]
